Move AggregateOperationImpl to impl package
